# public/swagger.json
{
 "swagger": "2.0",
 "info": {
  "version": "0.2.0",
  "title": "Traits App API"
 },
 "host": "api-traits.herokuapp.com",
 "basePath": "/api",
 "schemes": [
  "http",
  "https"
 ],
 "consumes": [
  "application/json"
 ],
 "produces": [
  "application/json"
 ],
 "paths": {
  "/personalityTraits/{type}": {
   "put": {
    "tags": [
     "Personality Traits"
    ],
    "description": "add a new personality traits to personality",
    "parameters": [
     {
      "name": "type",
      "type": "string",
      "in": "path",
      "required": true
     },
     {
      "name": "personalitytrait",
      "description": "Personality traits properties",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/PersonalityTraits"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/userTraits/{username}": {
   "put": {
    "tags": [
     "User Traits"
    ],
    "description": "update a user trait",
    "parameters": [
     {
      "name": "username",
      "description": "Username",
      "type": "string",
      "in": "path",
      "required": true
     },
     {
      "name": "usertrait",
      "description": "User traits properties",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/UserTraits"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/story": {
   "post": {
    "tags": [
     "Story"
    ],
    "description": "add a new story",
    "parameters": [
     {
      "name": "type",
      "description": "story properties",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/Story"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/stories/{type}": {
   "get": {
    "tags": [
     "Story"
    ],
    "description": "get stories by type",
    "parameters": [
     {
      "name": "type",
      "type": "string",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GetDescriptionListResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/people": {
   "get": {
    "tags": [
     "People"
    ],
    "description": "get people list",
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GetPeopleListResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   },
   "post": {
    "tags": [
     "People"
    ],
    "description": "add a new person to the list",
    "parameters": [
     {
      "name": "type",
      "description": "People properties",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/People"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/people/{type}": {
   "get": {
    "tags": [
     "People"
    ],
    "description": "get people by type",
    "parameters": [
     {
      "name": "type",
      "type": "string",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GetPeopleListResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/personality": {
   "get": {
    "tags": [
     "Personality"
    ],
    "description": "get personality list",
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GetPersonalitysListResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   },
   "post": {
    "tags": [
     "Personality"
    ],
    "description": "add a new personality to the list",
    "parameters": [
     {
      "name": "type",
      "description": "Personality properties",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/Personality"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/personality/{type}": {
   "get": {
    "tags": [
     "Personality"
    ],
    "description": "get a personality",
    "parameters": [
     {
      "name": "type",
      "type": "string",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GetPersonalityResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   },
   "put": {
    "tags": [
     "Personality"
    ],
    "description": "update a personality",
    "parameters": [
     {
      "name": "type",
      "description": "Personality type",
      "type": "string",
      "in": "path",
      "required": true
     },
     {
      "name": "personality",
      "description": "Personality properties",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/Personality"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Personality"
    ],
    "description": "delete a personality",
    "parameters": [
     {
      "name": "type",
      "description": "Personality type",
      "type": "string",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/userTypePart/{username}": {
   "get": {
    "tags": [
     "User Type Part"
    ],
    "description": "get a user",
    "parameters": [
     {
      "name": "username",
      "type": "string",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   },
   "put": {
    "tags": [
     "User Type Part"
    ],
    "description": "update a user type part",
    "parameters": [
     {
      "name": "username",
      "description": "username",
      "type": "string",
      "in": "path",
      "required": true
     },
     {
      "name": "usertypepart",
      "description": "User type part properties",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/UserTypePart"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   },
   "delete": {
    "tags": [
     "User Type Part"
    ],
    "description": "delete a movie",
    "parameters": [
     {
      "name": "username",
      "description": "Username",
      "type": "string",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/user": {
   "post": {
    "tags": [
     "User"
    ],
    "description": "add a new user to the list",
    "parameters": [
     {
      "name": "email",
      "description": "User properties",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/User"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GetUserResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/usersignin": {
   "post": {
    "tags": [
     "User"
    ],
    "description": "add a new user to the list",
    "parameters": [
     {
      "name": "email",
      "description": "User properties",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/UserSignin"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GetUserResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/user/{username}": {
   "get": {
    "tags": [
     "User"
    ],
    "description": "get a user",
    "parameters": [
     {
      "name": "username",
      "type": "string",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GetUserResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   },
   "patch": {
    "tags": [
     "User"
    ],
    "description": "update a user",
    "parameters": [
     {
      "name": "username",
      "description": "Username",
      "type": "string",
      "in": "path",
      "required": true
     },
     {
      "name": "user",
      "description": "user properties",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/UpdateUser"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   },
   "delete": {
    "tags": [
     "User"
    ],
    "description": "delete a user",
    "parameters": [
     {
      "name": "username",
      "description": "User id",
      "type": "string",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "Success",
      "schema": {
       "$ref": "#/definitions/GeneralResponse"
      }
     },
     "default": {
      "description": "Error",
      "schema": {
       "$ref": "#/definitions/ErrorResponse"
      }
     }
    }
   }
  },
  "/swagger": {
   "x-swagger-pipe": "swagger_raw"
  }
 },
 "definitions": {
  "GetDescriptionListResponse": {
   "required": [
    "descriptions"
   ],
   "properties": {
    "descriptions": {
     "type": "array",
     "items": {
      "type": "object",
      "properties": {
       "title": {
        "type": "string"
       },
       "description": {
        "type": "string"
       },
       "tags": {
        "type": "string"
       },
       "source": {
        "type": "string"
       },
       "timestamp": {
        "type": "string"
       },
       "personalitytype": {
        "type": "string"
       }
      }
     }
    }
   }
  },
  "GetPeopleListResponse": {
   "required": [
    "people_list"
   ],
   "properties": {
    "people_list": {
     "type": "array",
     "items": {
      "type": "object",
      "properties": {
       "id": {
        "type": "string"
       },
       "name": {
        "type": "string"
       },
       "lastname": {
        "type": "string"
       },
       "occupation": {
        "type": "string"
       },
       "isreal": {
        "type": "string"
       },
       "from": {
        "type": "string"
       },
       "personalitytype": {
        "type": "string"
       }
      }
     }
    }
   }
  },
  "GetPersonalitysListResponse": {
   "required": [
    "personalities"
   ],
   "properties": {
    "personalities": {
     "type": "array",
     "items": {
      "type": "object",
      "properties": {
       "id": {
        "type": "string"
       },
       "type": {
        "type": "string"
       },
       "totalfrequency": {
        "type": "string"
       },
       "malefrequency": {
        "type": "string"
       },
       "femalefrequency": {
        "type": "string"
       },
       "description": {
        "type": "string"
       },
       "typefull": {
        "type": "string"
       },
       "color": {
        "type": "string"
       }
      }
     }
    }
   }
  },
  "GetUserResponse": {
   "required": [
    "email"
   ],
   "properties": {
    "id": {
     "type": "string"
    },
    "name": {
     "type": "string"
    },
    "lastname": {
     "type": "string"
    },
    "email": {
     "type": "string"
    },
    "username": {
     "type": "string"
    },
    "passwordHash": {
     "type": "string"
    },
    "salt": {
     "type": "string"
    },
    "usertypeparts": {
     "type": "array",
     "items": {
      "$ref": "#/definitions/TypePart"
     }
    },
    "usertraits": {
     "type": "array",
     "items": {
      "$ref": "#/definitions/UserTraits"
     }
    }
   }
  },
  "GetPersonalityResponse": {
   "required": [
    "type"
   ],
   "properties": {
    "id": {
     "type": "string"
    },
    "type": {
     "type": "string"
    },
    "totalfrequency": {
     "type": "string"
    },
    "malefrequency": {
     "type": "string"
    },
    "femalefrequency": {
     "type": "string"
    },
    "description": {
     "type": "string"
    },
    "typefull": {
     "type": "string"
    },
    "color": {
     "type": "string"
    }
   }
  },
  "UserTypePart": {
   "type": "object",
   "properties": {
    "type_percentage": {
     "type": "array",
     "items": {
      "$ref": "#/definitions/TypePart"
     }
    }
   }
  },
  "TypePart": {
   "type": "object",
   "properties": {
    "personalitytype": {
     "type": "string",
     "description": "Personality type"
    },
    "percentage": {
     "type": "string",
     "description": "Personality percentage"
    },
    "lastupdate": {
     "type": "string",
     "format": "date-time",
     "description": "Last update. E.g. '2016-06-17T15:28:34Z' (RFC 3339, ISO 8601)"
    }
   }
  },
  "UserTraits": {
   "type": "object",
   "properties": {
    "personalitytype": {
     "type": "string",
     "description": "personality type for user selected traits"
    },
    "traits": {
     "type": "array",
     "items": {
      "type": "string"
     }
    }
   }
  },
  "PersonalityTraits": {
   "type": "object",
   "properties": {
    "traits": {
     "type": "array",
     "items": {
      "type": "string"
     }
    }
   }
  },
  "Trait": {
   "type": "object",
   "properties": {
    "trait": {
     "type": "string",
     "description": "Personality trait"
    },
    "weight": {
     "type": "number",
     "description": "weight determined by how often the trait is used for the personality type"
    }
   }
  },
  "User": {
   "type": "object",
   "properties": {
    "email": {
     "type": "string",
     "description": "task object name"
    },
    "password": {
     "type": "string",
     "description": "task object password"
    },
    "name": {
     "type": "string",
     "description": "user name"
    },
    "lastname": {
     "type": "string",
     "description": "user lastname"
    },
    "timestamp": {
     "type": "string",
     "format": "date-time",
     "description": "The date the user registered (E.g. '2016-06-17T15:28:34Z' (RFC 3339, ISO 8601))"
    },
    "usertraits": {
     "type": "array",
     "items": {
      "$ref": "#/definitions/UserTraits"
     }
    }
   },
   "required": [
    "email",
    "password"
   ]
  },
  "UpdateUser": {
   "type": "object",
   "properties": {
    "name": {
     "type": "string",
     "description": "user name"
    },
    "lastname": {
     "type": "string",
     "description": "user lastname"
    }
   }
  },
  "UserSignin": {
   "type": "object",
   "properties": {
    "email": {
     "type": "string",
     "description": "user email"
    },
    "password": {
     "type": "string",
     "description": "user password"
    }
   }
  },
  "Personality": {
   "type": "object",
   "properties": {
    "type": {
     "type": "string",
     "description": "task object name"
    },
    "totalfrequency": {
     "type": "string",
     "description": "task description"
    },
    "malefrequency": {
     "type": "string",
     "description": "task description"
    },
    "femalefrequency": {
     "type": "string",
     "description": "task description"
    },
    "description": {
     "type": "string",
     "description": "task description"
    },
    "typefull": {
     "type": "string",
     "description": "task description"
    },
    "color": {
     "type": "string",
     "description": "task description"
    },
    "traits": {
     "type": "array",
     "items": {
      "$ref": "#/definitions/Trait"
     }
    }
   },
   "required": [
    "type"
   ]
  },
  "People": {
   "type": "object",
   "properties": {
    "name": {
     "type": "string"
    },
    "lastname": {
     "type": "string"
    },
    "occupation": {
     "type": "string"
    },
    "isreal": {
     "type": "string"
    },
    "info": {
     "type": "string"
    },
    "personalitytype": {
     "type": "string"
    },
    "tags": {
     "type": "array",
     "items": {
      "type": "string"
     }
    }
   },
   "required": [
    "name"
   ]
  },
  "Story": {
   "type": "object",
   "properties": {
    "title": {
     "type": "string"
    },
    "description": {
     "type": "string"
    },
    "source": {
     "type": "string"
    },
    "timestamp": {
     "type": "string"
    },
    "personalitytype": {
     "type": "string"
    },
    "tags": {
     "type": "array",
     "items": {
      "type": "string"
     }
    }
   },
   "required": [
    "title"
   ]
  },
  "GeneralResponse": {
   "type": "object",
   "properties": {
    "success": {
     "type": "number",
     "description": "returns 1 if successful"
    },
    "description": {
     "type": "string",
     "description": "a short comment"
    }
   },
   "required": [
    "success",
    "description"
   ]
  },
  "ErrorResponse": {
   "required": [
    "message"
   ],
   "properties": {
    "message": {
     "type": "string"
    }
   }
  }
 }
}